마늘 무기 구현

+++ Assets/02.Scripts/MeleeWeapon/MeleeWeapon3Bullet.cs
@@ -4,11 +4,10 @@ using UnityEngine;
 
 public class MeleeWeapon3Bullet : MonoBehaviour
 {
-    public float baseDamage = 10f;
+     public int baseDamage = 10;
 
     void OnTriggerEnter2D(Collider2D hitInfo)
     {
-        // 각 적들마다 알아서 데미지 주기
         BaseEnemy enemy = hitInfo.GetComponent<BaseEnemy>();
         if (enemy != null)
         {
